TileComponent

+++ GameEngine/GameEngine/TransformComponent.h
@@ -28,7 +28,7 @@ public:
 		position.y = y;
 	}
 
-	TransformComponent(float x, int y, int h, int w, int sc) {
+	TransformComponent(float x, float y, int h, int w, int sc) {
 		position.x = x;
 		position.y = y;
 		height = h;
